+++ front/src/app/(views)/visitas/components/InfoBox.tsx
